DO NOT MERGE. (read description)
not sure if this will actually run since it's an sql file that needs to be on a server. but I put my code here so you can peruse at your leisure. maybe it'll help you understand my db a little better. you can look at the txt file or the sql file. ensure you add the PHP intelephense package and SQLTools package to vs code for easier reading/